requirements.txt
fastui[fastapi]
uvicorn
pydantic
llama-index>=0.11.0
llama-index-readers-file
llama-index-vector-stores-postgres
llama-index-embeddings-openai
llama-index-embeddings-ollama
llama-index-embeddings-bedrock
llama-index-llms-litellm
PyYAML>=6.0
rich
python-dotenv
fastapi
psycopg2-binary
python-multipart
pypdf
python-docx